[pylint] Rename 'too-many-args' to 'too-many-arguments'
The actual name
